fix: raycast calc
