+++ src/render/raycast.c
@@ -65,7 +65,7 @@ void	calc_step_and_side_dist(t_ray *ray, t_player *player)
 	else
 	{
 		ray->step_x = 1;
-		ray->side_dist_x = (ray->map_x * 1.0 - player->x) * ray->delta_dist_x;
+		ray->side_dist_x = (ray->map_x + 1.0 - player->x) * ray->delta_dist_x;
 	}
 	if (ray->dir_y < 0)
 	{
@@ -75,6 +75,6 @@ void	calc_step_and_side_dist(t_ray *ray, t_player *player)
 	else
 	{
 		ray->step_y = 1;
-		ray->side_dist_y = (ray->map_y * 1.0 - player->y) * ray->delta_dist_y;
+		ray->side_dist_y = (ray->map_y + 1.0 - player->y) * ray->delta_dist_y;
 	}
 }
